Rename package bootstrap -> job
