+++ hook/scriptwrapper.go
@@ -10,10 +10,10 @@ import (
 	"strings"
 	"text/template"
 
-	"github.com/buildkite/agent/v3/bootstrap/shell"
 	"github.com/buildkite/agent/v3/env"
 	"github.com/buildkite/agent/v3/internal/shellscript"
 	"github.com/buildkite/agent/v3/internal/utils"
+	"github.com/buildkite/agent/v3/job/shell"
 )
 
 const (
