+++ internal/redactor/redactor.go
@@ -7,7 +7,7 @@ import (
 	"path"
 	"sync"
 
-	"github.com/buildkite/agent/v3/bootstrap/shell"
+	"github.com/buildkite/agent/v3/job/shell"
 )
 
 // RedactLengthMin is the shortest string length that will be considered a
