+++ job/bootstrap_test.go
@@ -1,11 +1,11 @@
-package bootstrap
+package job
 
 import (
 	"context"
 	"testing"
 
-	"github.com/buildkite/agent/v3/bootstrap/shell"
 	"github.com/buildkite/agent/v3/internal/redactor"
+	"github.com/buildkite/agent/v3/job/shell"
 	"github.com/buildkite/agent/v3/tracetools"
 	"github.com/google/go-cmp/cmp"
 	"github.com/opentracing/opentracing-go"
