+++ job/docker.go
@@ -1,4 +1,4 @@
-package bootstrap
+package job
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/buildkite/agent/v3/bootstrap/shell"
+	"github.com/buildkite/agent/v3/job/shell"
 )
 
 var dockerEnv = []string{
