+++ job/knownhosts_test.go
@@ -1,4 +1,4 @@
-package bootstrap
+package job
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/buildkite/agent/v3/bootstrap/shell"
+	"github.com/buildkite/agent/v3/job/shell"
 	"github.com/gliderlabs/ssh"
 )
