+++ job/shell/lookpath.go
